refactor: refined the way realm classes were implemented
